Create LittleTacoShop.html for website structure
This is my first commit. This is the first change, no real "change" here as this is the first bit of code I copy pasted into my GitHub repository from VSCode. Realized it is more organized by writing the actual code inside Github to see version history and do necessary reverts if needed.